Refactored new ToolInput/Outputs so they reside in own packages and not in original interfaces
