+++ src/org/openjdk/asmtools/common/inputs/StdinInput.java
@@ -1,0 +1,14 @@
+package org.openjdk.asmtools.common.inputs;
+
+public class StdinInput extends StreamInput {
+
+    public StdinInput() {
+        super(System.in);
+    }
+
+    @Override
+    public String getFileName() {
+        //get parent is used
+        return "stdin/stdin";
+    }
+}
