+++ src/org/openjdk/asmtools/common/outputs/log/SingleDualOutputStreamOutput.java
@@ -1,0 +1,18 @@
+package org.openjdk.asmtools.common.outputs.log;
+
+import java.io.PrintStream;
+import java.io.PrintWriter;
+
+public class SingleDualOutputStreamOutput extends DualOutputStreamOutput {
+    public SingleDualOutputStreamOutput() {
+        this(System.err);
+    }
+
+    public SingleDualOutputStreamOutput(PrintWriter er) {
+        super(er, er);
+    }
+
+    public SingleDualOutputStreamOutput(PrintStream er) {
+        super(er, er);
+    }
+}
