+++ src/org/openjdk/asmtools/jdis/Indenter.java
@@ -22,10 +22,9 @@
  */
 package org.openjdk.asmtools.jdis;
 
-import org.openjdk.asmtools.common.ToolOutput;
+import org.openjdk.asmtools.common.outputs.ToolOutput;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Formatter;
