+++ test/java/org/openjdk/asmtools/jasm/MultipleFilesInSingleJasmTest.java
@@ -2,8 +2,11 @@ package org.openjdk.asmtools.jasm;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.openjdk.asmtools.common.ToolInput;
-import org.openjdk.asmtools.common.ToolOutput;
+import org.openjdk.asmtools.common.inputs.ByteInput;
+import org.openjdk.asmtools.common.inputs.ToolInput;
+import org.openjdk.asmtools.common.outputs.ByteOutput;
+import org.openjdk.asmtools.common.outputs.log.DualStreamToolOutput;
+import org.openjdk.asmtools.common.outputs.log.SingleDualOutputStreamOutput;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -14,9 +17,9 @@ public class MultipleFilesInSingleJasmTest {
     @Test
     public void clfacc00610m10pTest() throws IOException {
         byte[] jasmFile = getJasmFile("clfacc00610m10p.jasm");
-        ToolInput file = new ToolInput.ByteInput(jasmFile);
-        ToolOutput.ByteOutput output = new ToolOutput.ByteOutput();
-        ToolOutput.DualStreamToolOutput log = new ToolOutput.SingleDualOutputStreamOutput(); //todo hide to ToolOutput.StringLog once done
+        ToolInput file = new ByteInput(jasmFile);
+        ByteOutput output = new ByteOutput();
+        DualStreamToolOutput log = new SingleDualOutputStreamOutput(); //todo hide to ToolOutput.StringLog once done
         org.openjdk.asmtools.jasm.Main jasm = new org.openjdk.asmtools.jasm.Main(output, log, file, "-v");
         int i = jasm.compile();
         Assertions.assertEquals(0, i);
@@ -38,9 +41,9 @@ public class MultipleFilesInSingleJasmTest {
     @Test
     public void spinum00101m10pTest() throws IOException {
         byte[] jasmFile = getJasmFile("spinum00101m10p.jasm");
-        ToolInput file = new ToolInput.ByteInput(jasmFile);
-        ToolOutput.ByteOutput output = new ToolOutput.ByteOutput();
-        ToolOutput.DualStreamToolOutput log = new ToolOutput.SingleDualOutputStreamOutput(); //todo hide to ToolOutput.StringLog once done
+        ToolInput file = new ByteInput(jasmFile);
+        ByteOutput output = new ByteOutput();
+        DualStreamToolOutput log = new SingleDualOutputStreamOutput(); //todo hide to ToolOutput.StringLog once done
         org.openjdk.asmtools.jasm.Main jasm = new org.openjdk.asmtools.jasm.Main(output, log, file, "-v");
         int i = jasm.compile();
         Assertions.assertEquals(0, i);
