+++ test/java/org/openjdk/asmtools/jcoder/MultipleFilesInSingleJcoderTest.java
@@ -2,8 +2,11 @@ package org.openjdk.asmtools.jcoder;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.openjdk.asmtools.common.ToolInput;
-import org.openjdk.asmtools.common.ToolOutput;
+import org.openjdk.asmtools.common.inputs.ByteInput;
+import org.openjdk.asmtools.common.inputs.ToolInput;
+import org.openjdk.asmtools.common.outputs.ByteOutput;
+import org.openjdk.asmtools.common.outputs.log.DualStreamToolOutput;
+import org.openjdk.asmtools.common.outputs.log.SingleDualOutputStreamOutput;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -14,9 +17,9 @@ public class MultipleFilesInSingleJcoderTest {
     @Test
     public void jcod12Test() throws IOException {
         byte[] jcodFile = getJcodFile("12.jcod");
-        ToolInput file = new ToolInput.ByteInput(jcodFile);
-        ToolOutput.ByteOutput output = new ToolOutput.ByteOutput();
-        ToolOutput.DualStreamToolOutput log = new ToolOutput.SingleDualOutputStreamOutput(); //todo hide to ToolOutput.StringLog once done
+        ToolInput file = new ByteInput(jcodFile);
+        ByteOutput output = new ByteOutput();
+        DualStreamToolOutput log = new SingleDualOutputStreamOutput(); //todo hide to ToolOutput.StringLog once done
         org.openjdk.asmtools.jcoder.Main jcod = new org.openjdk.asmtools.jcoder.Main(output, log, file, "-v");
         int i = jcod.compile();
         Assertions.assertEquals(0, i);
