+++ test/java/org/openjdk/asmtools/jdis/MainTest.java
@@ -5,7 +5,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.openjdk.asmtools.ClassPathClassWork;
 import org.openjdk.asmtools.ThreeStringWriters;
-import org.openjdk.asmtools.common.ToolOutput;
+import org.openjdk.asmtools.common.outputs.EscapedPrintStreamOutput;
+import org.openjdk.asmtools.common.outputs.log.SingleDualOutputStreamOutput;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -107,7 +108,7 @@ class MainTest extends ClassPathClassWork {
         dir.delete();
         dir.mkdir();
         dir.deleteOnExit();
-        org.openjdk.asmtools.jasm.Main jasmTool = new org.openjdk.asmtools.jasm.Main(new ToolOutput.EscapedPrintStreamOutput(System.out), new ToolOutput.SingleDualOutputStreamOutput(), sourceWithoutSuper.getAbsolutePath(), "-d", dir.getAbsolutePath());
+        org.openjdk.asmtools.jasm.Main jasmTool = new org.openjdk.asmtools.jasm.Main(new EscapedPrintStreamOutput(System.out), new SingleDualOutputStreamOutput(), sourceWithoutSuper.getAbsolutePath(), "-d", dir.getAbsolutePath());
         int ii = jasmTool.compile();
         Assertions.assertEquals(0, ii);
         ThreeStringWriters outs = new ThreeStringWriters();
